Separate the dict_validate_lib's unit tests into several atomic tests
