--- fuzzing/tools/dict_data/invalid.dict
@@ -1,3 +1,0 @@
-Invalid dictionary entries
-"
-"\A"
